server/requirements.txt
annotated-types==0.7.0
anyio==4.4.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
bcrypt==4.2.0
cachetools==5.4.0
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==43.0.0
dnspython==2.6.1
email_validator==2.1.2
fastapi==0.111.1
fastapi-cli==0.0.4
google-ai-generativelanguage==0.6.6
google-api-core==2.19.1
google-api-python-client==2.138.0
google-auth==2.32.0
google-auth-httplib2==0.2.0
google-generativeai==0.7.2
googleapis-common-protos==1.63.2
greenlet==3.0.3
grpcio==1.65.1
grpcio-status==1.62.2
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httptools==0.6.1
httpx==0.27.0
idna==3.7
Jinja2==3.1.4
makefun==1.15.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
pillow==10.4.0
proto-plus==1.24.0
protobuf==4.25.4
psycopg2-binary==2.9.9
pwdlib==0.2.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
PyJWT==2.8.0
pyparsing==3.1.2
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.1
requests==2.32.3
rich==13.7.1
rsa==4.9
setuptools==69.5.1
shellingham==1.5.4
sniffio==1.3.1
SQLAlchemy==2.0.31
sqlalchemy-cockroachdb==2.0.2
starlette==0.37.2
tqdm==4.66.4
typer==0.12.3
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.2
uvicorn==0.30.3
uvloop==0.19.0
watchfiles==0.22.0
websockets==12.0
wheel==0.43.0
pymongo==4.8.0
pytest==8.3.2 
# Sphinx==8.0.2 
# sphinx-rtd-theme==2.0.0 
# sphinxcontrib-httpdomain==1.8.1 